+++ sqlg-core/src/main/java/org/umlg/sqlg/structure/TopologyListener.java
@@ -6,5 +6,5 @@ package org.umlg.sqlg.structure;
  */
 public interface TopologyListener {
 
-    void change(TopologyInf topologyInf, String oldValue, TopologyChangeAction action);
+    void change(TopologyInf topologyInf, TopologyInf oldValue, TopologyChangeAction action);
 }