+++ sqlg-core/src/main/java/org/umlg/sqlg/structure/topology/Partition.java
@@ -227,7 +227,7 @@ public class Partition implements TopologyInf {
             } else {
                 partition.detach();
             }
-            this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.DELETE);
+            this.getAbstractLabel().getSchema().getTopology().fire(partition, partition, TopologyChangeAction.DELETE);
         }
     }
 
@@ -1073,7 +1073,7 @@ public class Partition implements TopologyInf {
                 String pName = jsonNode.asText();
                 Partition old = this.partitions.remove(pName);
                 if (fire && old != null) {
-                    this.getAbstractLabel().getSchema().getTopology().fire(old, "", TopologyChangeAction.DELETE);
+                    this.getAbstractLabel().getSchema().getTopology().fire(old, old, TopologyChangeAction.DELETE);
                 }
             }
         }
@@ -1172,7 +1172,7 @@ public class Partition implements TopologyInf {
         this.uncommittedPartitions.remove(name);
         Partition partition = Partition.createRangeSubPartition(this.sqlgGraph, this, name, from, to);
         this.uncommittedPartitions.put(name, partition);
-        this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.CREATE);
+        this.getAbstractLabel().getSchema().getTopology().fire(partition, null, TopologyChangeAction.CREATE);
         return partition;
     }
 
@@ -1181,7 +1181,7 @@ public class Partition implements TopologyInf {
         this.uncommittedPartitions.remove(name);
         Partition partition = Partition.createListSubPartition(this.sqlgGraph, this, name, in);
         this.uncommittedPartitions.put(name, partition);
-        this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.CREATE);
+        this.getAbstractLabel().getSchema().getTopology().fire(partition, null, TopologyChangeAction.CREATE);
         return partition;
     }
 
@@ -1190,7 +1190,7 @@ public class Partition implements TopologyInf {
         this.uncommittedPartitions.remove(name);
         Partition partition = Partition.createHashSubPartition(this.sqlgGraph, this, name, modulus, remainder);
         this.uncommittedPartitions.put(name, partition);
-        this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.CREATE);
+        this.getAbstractLabel().getSchema().getTopology().fire(partition, null, TopologyChangeAction.CREATE);
         return partition;
     }
 
@@ -1212,7 +1212,7 @@ public class Partition implements TopologyInf {
                 partitionType,
                 partitionExpression);
         this.uncommittedPartitions.put(name, partition);
-        this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.CREATE);
+        this.getAbstractLabel().getSchema().getTopology().fire(partition, null, TopologyChangeAction.CREATE);
         return partition;
     }
 
@@ -1232,7 +1232,7 @@ public class Partition implements TopologyInf {
                 partitionType,
                 partitionExpression);
         this.uncommittedPartitions.put(name, partition);
-        this.getAbstractLabel().getSchema().getTopology().fire(partition, "", TopologyChangeAction.CREATE);
+        this.getAbstractLabel().getSchema().getTopology().fire(partition, null, TopologyChangeAction.CREATE);
         return partition;
     }