+++ sqlg-test/src/main/java/org/umlg/sqlg/test/topology/TestTopologyChangeListener.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.*;
  */
 public class TestTopologyChangeListener extends BaseTest {
 
-    private final List<Triple<TopologyInf, String, TopologyChangeAction>> topologyListenerTriple = new ArrayList<>();
+    private final List<Triple<TopologyInf, TopologyInf, TopologyChangeAction>> topologyListenerTriple = new ArrayList<>();
 
     @Before
     public void before() throws Exception {
@@ -56,7 +56,7 @@ public class TestTopologyChangeListener extends BaseTest {
         assertEquals(8, this.topologyListenerTriple.size());
 
         assertEquals(schema, this.topologyListenerTriple.get(0).getLeft());
-        assertEquals("", this.topologyListenerTriple.get(0).getMiddle());
+        assertNull(this.topologyListenerTriple.get(0).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(0).getRight());
 
         assertEquals(aVertexLabel, this.topologyListenerTriple.get(1).getLeft());
@@ -64,7 +64,7 @@ public class TestTopologyChangeListener extends BaseTest {
         assertTrue(props.containsKey("name"));
         assertTrue(props.containsKey("surname"));
 
-        assertEquals("", this.topologyListenerTriple.get(1).getMiddle());
+        assertNull(this.topologyListenerTriple.get(1).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(1).getRight());
 
         assertEquals(edgeLabel, this.topologyListenerTriple.get(2).getLeft());
@@ -72,36 +72,36 @@ public class TestTopologyChangeListener extends BaseTest {
         assertTrue(s.contains(edgeLabel.getSchema().getName()));
         props = ((EdgeLabel) this.topologyListenerTriple.get(2).getLeft()).getProperties();
         assertTrue(props.containsKey("special"));
-        assertEquals("", this.topologyListenerTriple.get(2).getMiddle());
+        assertNull(this.topologyListenerTriple.get(2).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(2).getRight());
 
         assertEquals(vertexPropertyColumn, this.topologyListenerTriple.get(3).getLeft());
-        assertEquals("", this.topologyListenerTriple.get(3).getMiddle());
+        assertNull(this.topologyListenerTriple.get(3).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(3).getRight());
 
         assertEquals(edgePropertyColumn, this.topologyListenerTriple.get(4).getLeft());
-        assertEquals("", this.topologyListenerTriple.get(4).getMiddle());
+        assertNull(this.topologyListenerTriple.get(4).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(4).getRight());
 
         assertEquals(bVertexLabel, this.topologyListenerTriple.get(5).getLeft());
-        assertEquals("", this.topologyListenerTriple.get(5).getMiddle());
+        assertNull(this.topologyListenerTriple.get(5).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(5).getRight());
 
         assertEquals(edgeLabel, this.topologyListenerTriple.get(6).getLeft());
-        assertEquals("A.B", this.topologyListenerTriple.get(6).getMiddle());
+        assertEquals(bVertexLabel, this.topologyListenerTriple.get(6).getMiddle());
         assertEquals(TopologyChangeAction.ADD_IN_VERTEX_LABELTO_EDGE, this.topologyListenerTriple.get(6).getRight());
 
         assertEquals(index, this.topologyListenerTriple.get(7).getLeft());
-        assertEquals("", this.topologyListenerTriple.get(7).getMiddle());
+        assertNull(this.topologyListenerTriple.get(7).getMiddle());
         assertEquals(TopologyChangeAction.CREATE, this.topologyListenerTriple.get(7).getRight());
 
         this.sqlgGraph.tx().commit();
     }
 
     public static class TopologyListenerTest implements TopologyListener {
-        private List<Triple<TopologyInf, String, TopologyChangeAction>> topologyListenerTriple = new ArrayList<>();
+        private List<Triple<TopologyInf, TopologyInf, TopologyChangeAction>> topologyListenerTriple = new ArrayList<>();
 
-        public TopologyListenerTest(List<Triple<TopologyInf, String, TopologyChangeAction>> topologyListenerTriple) {
+        public TopologyListenerTest(List<Triple<TopologyInf, TopologyInf, TopologyChangeAction>> topologyListenerTriple) {
             super();
             this.topologyListenerTriple = topologyListenerTriple;
         }
@@ -111,7 +111,7 @@ public class TestTopologyChangeListener extends BaseTest {
         }
 
         @Override
-        public void change(TopologyInf topologyInf, String oldValue, TopologyChangeAction action) {
+        public void change(TopologyInf topologyInf, TopologyInf oldValue, TopologyChangeAction action) {
             String s = topologyInf.toString();
             assertNotNull(s);
             assertTrue(s + "does not contain " + topologyInf.getName(), s.contains(topologyInf.getName()));
@@ -121,7 +121,7 @@ public class TestTopologyChangeListener extends BaseTest {
         }
 
         public boolean receivedEvent(TopologyInf topologyInf, TopologyChangeAction action) {
-            for (Triple<TopologyInf, String, TopologyChangeAction> t : topologyListenerTriple) {
+            for (Triple<TopologyInf, TopologyInf, TopologyChangeAction> t : topologyListenerTriple) {
                 if (t.getLeft().equals(topologyInf) && t.getRight().equals(action)) {
                     return true;
                 }